+++ roulette-app/wallet-service/src/main/java/com/jfCasino/wallet_service/dto/request/WalletCommitRequest.java
@@ -5,6 +5,7 @@ import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.Min;
 import jakarta.validation.constraints.NotBlank;
 import java.util.UUID;
+import java.math.BigDecimal;
 
 @Schema(
     name = "WalletCommitRequest",
@@ -31,7 +32,7 @@ public class WalletCommitRequest {
         minimum = "0"
     )
     @Min(0)
-    private int amount;
+    private BigDecimal amount;
     public WalletCommitRequest() {};
 
     //getters and setters
@@ -51,11 +52,11 @@ public class WalletCommitRequest {
         this.userID = userID;
     }
 
-    public int getAmount() {
+    public BigDecimal getAmount() {
         return amount;
     }
 
-    public void setAmount(int amount) {
+    public void setAmount(BigDecimal amount) {
         this.amount = amount;
     }
 }
